# perf: use path aliases and swc for test:perf tasks to improve build speeds