--- a/azure-pipelines.perf-test.yml
+++ b/azure-pipelines.perf-test.yml
@@ -36,7 +36,7 @@ jobs:
         name: PackagesAffected
         displayName: Check if v8, v9 and/or northstar packages were affected
 
-  - job: PerfTestNorthstar
+  - job: PerfTestReactNorthstar
     displayName: Perf test for @fluentui/react-northstar
     dependsOn: CheckIfPackagesAffected
     condition: eq(dependencies.CheckIfPackagesAffected.outputs['PackagesAffected.NorthstarPackageAffected'], true)
@@ -50,21 +50,20 @@ jobs:
           filePath: yarn-ci.sh
         displayName: yarn
 
-      - script: |
-          yarn lage build --to @fluentui/perf-test-northstar --to @fluentui/react-conformance --verbose
-        condition: eq(variables.isPR, true)
-        displayName: Build to Perf Test (Northstar)
+      # - script: |
+      #     # yarn lage build --to @fluentui/perf-test-northstar --to @fluentui/react-conformance --verbose
+      #     yarn workspace @fluentui/digest build
+      #   condition: eq(variables.isPR, true)
+      #   displayName: Build to Perf Test (Northstar)
 
       - script: |
-          yarn perf:test --base
+          yarn workspace @fluentui/perf-test-northstar perf:test --base
         condition: eq(variables.isPR, false)
-        workingDirectory: packages/fluentui/perf-test-northstar
         displayName: Run Perf Test Base (Northstar)
 
       - script: |
-          yarn perf:test
+          yarn workspace @fluentui/perf-test-northstar perf:test
         condition: eq(variables.isPR, true)
-        workingDirectory: packages/fluentui/perf-test-northstar
         displayName: Run Perf Test (Northstar)
 
       - task: AzureUpload@2