--- a/packages/fluentui/digest/package.json
+++ b/packages/fluentui/digest/package.json
@@ -5,9 +5,6 @@
   "main": "lib/digest.js",
   "module": "lib/digest.js",
   "license": "MIT",
-  "bin": {
-    "digest": "./bin/digest.js"
-  },
   "repository": {
     "type": "git",
     "url": "https://github.com/microsoft/digest"