--- a/packages/fluentui/perf-test-northstar/tasks/fluentPerfRegressions.ts
+++ b/packages/fluentui/perf-test-northstar/tasks/fluentPerfRegressions.ts
@@ -1,6 +1,6 @@
 import * as _ from 'lodash';
 import * as path from 'path';
-import { workspaceRoot } from 'nx/src/utils/app-root';
+import { workspaceRoot } from '@nrwl/devkit';
 import { perfTestEnv } from '@fluentui/scripts-tasks';
 
 import { config } from './perf-test.config';