--- a/scripts/storybook/src/rules.js
+++ b/scripts/storybook/src/rules.js
@@ -1,6 +1,42 @@
 const { _createCodesandboxRule } = require('./utils');
 
 /**
+ * @type {import("webpack").RuleSetRule}
+ */
+const tsRule = {
+  test: [/\.tsx?$/],
+  use: {
+    loader: 'swc-loader',
+    options: {
+      jsc: {
+        target: 'es2015',
+        parser: {
+          syntax: 'typescript',
+          tsx: true,
+          decorators: true,
+          dynamicImport: true,
+        },
+        transform: {
+          decoratorMetadata: true,
+          legacyDecorator: true,
+        },
+        keepClassNames: true,
+        externalHelpers: true,
+        loose: true,
+      },
+    },
+  },
+};
+
+/**
+ * @type {import("webpack").RuleSetRule}
+ */
+const cssRule = {
+  test: /\.css$/,
+  use: ['style-loader', 'css-loader'],
+};
+
+/**
  * v8 uses SCSS/CSS modules
  * @type {import("webpack").RuleSetRule}
  */
@@ -61,6 +97,8 @@ const griffelRule = {
  */
 const codesandboxRule = _createCodesandboxRule();
 
+exports.tsRule = tsRule;
 exports.scssRule = scssRule;
+exports.cssRule = cssRule;
 exports.griffelRule = griffelRule;
 exports.codesandboxRule = codesandboxRule;