<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->



<table role="table">
  <thead>
    <tr>
      <th>Test suite</th>
      <th>Build time</th>
      <th>Delta</th>
      <th>Test time</th>
      <th>Delta</th>
    </tr>
  </thead>
  <tbody>
    <tr>
      <td>react-components: Before</td>
      <td>186s</td>
      <td rowspan="2">+96% 🚅</td>
      <td>163s</td>
      <td rowspan="2"><del>-63%</del>/ after rebase 0%</td>
    </tr>
    <tr>
      <td>react-components: After</td>
      <td>8s</td>
      <td><del>262s</del> / after rebase 162s</td>
    </tr>
    <tr>
      <td>react: Before</td>
      <td>134s</td>
      <td rowspan="2">+92% 🚅</td>
      <td>1231s</td>
      <td rowspan="2"><del>-65%</del> / after rebase ~0%</td>
    </tr>
    <tr>
      <td>react: After</td>
      <td>11s</td>
      <td><del>1800s</del> / after rebase 1227s</td>
    </tr>
    <tr>
      <td>react-northstar: Before</td>
      <td>490s</td>
      <td rowspan="2" style="background-color: greenyellow;">+93% 🚅</td>
      <td>402s</td>
      <td rowspan="2" style="background-color: greenyellow;">+16.5% 🚅</td>
    </tr>
    <tr>
      <td>react-northstart: After</td>
      <td>35s</td>
      <td>345s</td>
    </tr>
  </tbody>
</table>

## Previous Behavior

<img width="1433" alt="image" src="https://user-images.githubusercontent.com/1223799/219710589-6da5b9c1-1654-4c42-baab-61dea8b4b19b.png">

## New Behavior

<img width="1426" alt="image" src="https://user-images.githubusercontent.com/1223799/220085111-83d875fd-319b-403d-8450-dc3be0f9b1e6.png">

-  perf improvements
-  every perf app has unified `type-check` npm alias

## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Follows https://github.com/microsoft/fluentui/pull/26559